chore: add Env type definition
