+++ package.json
@@ -14,7 +14,7 @@
     "ai": "^5.0.22",
     "hono": "^4.4.7",
     "workers-ai-provider": "^0.7.5",
-    "zod": "^3.25.0"
+    "zod": "^3.25.76"
   },
   "devDependencies": {
     "@cloudflare/workers-types": "^4.20240512.0",
