+++ src/index.test.ts
@@ -5,7 +5,7 @@ vi.mock('ai', () => ({
 }));
 
 import app from './index';
-import type { Env } from './index';
+import type { Env } from './types';
 
 // Simple mocks for bindings with minimal type fixes
 const bindings: Env = {
