+++ src/lib/homeAssistantWs.ts
@@ -7,7 +7,7 @@
  * `env.HASSIO_ENDPOINT_URI`.
  */
 
-import type { Env } from '../index';
+import type { Env } from '../types';
 
 interface PendingRequest {
   resolve: (value: any) => void;
